Info on PostgreSQL brew upgrades on major versions
I have gotten caught by brew upgrading major versions of PostgreSQL, say
for example 11 to 12, 12 to 13, 13 to 14, etc.  When a major upgrade
occurs all DB's need to be "reformated".  While there are some
provisions for `brew upgrade` to handle this automatically, it has
failed for me, more times than it has worked.  I have added
information on how to to work around/with this behaviour.

Luckily, this behaviour should be a thing of the past.  `brew` now
includes the manjor version of PostgreSQL in the package name.  So now
when I install `postgresql@14`, there's no chance of version 15 be
automatically installed.